app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
requests
httpx
jinja2
python-multipart
beanie
motor
bcrypt
pyJWT
mailersend
redis
pydantic-settings